+++ src/rest/defaults.ts
@@ -1,14 +1,16 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
-import {ReportProgressMiddleware, ResponseTypeMiddleware} from './middlewares';
+import {ReportProgressMiddleware, ResponseTypeMiddleware, ResponseTransformMiddleware, IgnoredInterceptorsMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
  */
 export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
 [
+    ResponseTransformMiddleware,
     ResponseTypeMiddleware,
+    IgnoredInterceptorsMiddleware,
     ReportProgressMiddleware
 ];
 
@@ -17,6 +19,8 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
  */
 export const BASIC_DEFAULT_REST_METHOD_MIDDLEWARES: Type<RestMiddleware>[] =
 [
+    ResponseTransformMiddleware,
     ResponseTypeMiddleware,
+    IgnoredInterceptorsMiddleware,
     ReportProgressMiddleware
 ];