+++ src/rest/middlewares/index.ts
@@ -1,3 +1,4 @@
-// export * from './parameters.middleware';
 export * from './reportProgress.middleware';
 export * from './responseType.middleware';
+export * from './responseTransform.middleware';
+export * from './ignoredInterceptors.middleware';