+++ src/rest/rest.interface.ts
@@ -270,3 +270,24 @@ export interface RestMiddleware<TRequestBody = any, TResponseBody = any, TDescri
      */
     run: RestMiddlewareRunMethod<TRequestBody, TResponseBody, TDescriptor, TTarget>;
 }
+
+/**
+ * Defintion of buildMiddleware function type
+ */
+export interface BuildMiddlewaresFn
+{
+    /**
+     * Builds and returns array of middleware run functions
+     * @param middlewares - Array of set middleware types
+     * @param middlewaresOrder - Array of middleware types in order that should be executed
+     */
+    (middlewares: Type<RestMiddleware>[], middlewaresOrder: Type<RestMiddleware>[]): RestMiddlewareRunMethod[];
+
+    /**
+     * Builds and returns array of middleware run functions
+     * @param this - Instance of RESTClient
+     * @param middlewares - Array of set middleware types
+     * @param middlewaresOrder - Array of middleware types in order that should be executed
+     */
+    (this: ɵRestMethod, middlewares: Type<RestMiddleware>[], middlewaresOrder: Type<RestMiddleware>[]): RestMiddlewareRunMethod[]
+}